build: Factor out common code into core/build_util.rs (#8756)
