+++ core/lib.rs
@@ -8,6 +8,7 @@ extern crate log;
 mod async_cancel;
 mod async_cell;
 mod bindings;
+pub mod build_util;
 pub mod error;
 mod flags;
 mod gotham_state;
